# swagger-ui/api-docs.yaml
openapi: 3.0.3
info:
  title: Todo App REST API
  description: |
    A comprehensive REST API for a Todo application built with Express.js and JSON Server.
    
    ## Features
    - User registration and authentication
    - Todo CRUD operations
    - Category management
    - User-Todo relationships
    - JWT-based authentication
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
  version: 1.0.0
  contact:
    name: API Support
    email: support@todoapp.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /api-proxy
    description: Proxied API Server (Recommended for Codespaces)
  - url: http://localhost:5000
    description: Local Development Express Server
  - url: http://localhost:3001
    description: Local Development JSON Server
  - url: http://localhost:3001
    description: JSON Server (Database)
...
paths:
...
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Create a new user account with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - email
                - password
                - firstName
                - lastName
              properties:
                username:
                  type: string
                  minLength: 3
                  maxLength: 30
                  pattern: '^[a-zA-Z0-9_]+$'
                  example: "johndoe"
                email:
                  type: string
                  format: email
                  example: "john.doe@example.com"
                password:
                  type: string
                  minLength: 6
                  pattern: '^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)'
                  example: "Password123"
                firstName:
                  type: string
                  minLength: 1
                  maxLength: 50
                  example: "John"
                lastName:
                  type: string
                  minLength: 1
                  maxLength: 50
                  example: "Doe"
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error or user already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: Authenticate user with username/email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  description: Username or email address
                  example: "johndoe"
                password:
                  type: string
                  example: "Password123"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout user
      description: Logout the authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /api/auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user profile
      description: Get the profile of the authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
  /api/auth/change-password:
    post:
      tags:
        - Authentication
      summary: Change user password
      description: Change the password of the authenticated user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - currentPassword
                - newPassword
              properties:
                currentPassword:
                  type: string
                  example: "OldPassword123"
                newPassword:
                  type: string
                  minLength: 6
                  pattern: '^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)'
                  example: "NewPassword123"
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  # User Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: "1"
        username:
          type: string
          example: "johndoe"
        email:
          type: string
          format: email
          example: "john.doe@example.com"
        firstName:
          type: string
          example: "John"
        lastName:
          type: string
          example: "Doe"
        avatar:
          type: string
          example: ""
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
        loginAttempts:
          type: integer
          example: 0
        accountLockedUntil:
          type: string
          format: date-time
          nullable: true
        isActive:
          type: boolean
          example: true
        preferences:
          $ref: '#/components/schemas/UserPreferences'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserPreferences:
      type: object
      properties:
        theme:
          type: string
          enum: [light, dark, system]
          example: "light"
        language:
          type: string
          example: "en"
        notifications:
          type: object
          properties:
            browser:
              type: boolean
              example: true
            inApp:
              type: boolean
              example: true
        timezone:
          type: string
          example: "UTC"
...
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "User registered successfully"
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            token:
              type: string
              example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            message:
              type: string
              example: "Validation failed"
            details:
              type: array
              items:
                type: object